requirements.txt
# API y servidor
fastapi
uvicorn[standard]
python-multipart

asyncpg
SQLAlchemy
psycopg2-binary
 
# Claude
anthropic
 
# Utilidades
python-dotenv
rich
pandas
openpyxl
ortools
